AI players can now use customizable neural nets to think, however, I have not yet been able to find a perfect model for the game

+++ src/main/java/LabeledResult.java
@@ -1,9 +1,9 @@
 public class LabeledResult {
-  int[] features;
+  double[] features;
   int label;
   String id;
 
-  public LabeledResult(int[] features, int label, String id){
+  public LabeledResult(double[] features, int label, String id){
     this.features = features;
     this.label = label;
     this.id = id;
